Fix slider to prevent loss of color precision
Before this change, the color slider did a lossy conversion between
HSL <> RGB which resulted in loss of precision when the slider was taken
far to the left or right side.

This commit introduces variables fHSL and bHSL to keep track statically
of the HSL, so the hue and saturation doesn't change and lightness
becomes the only value that changes with the slider.